requirements.txt
langchain==0.0.162
openai==0.27.6
chromadb==0.3.22
python-dotenv~=1.0.0
streamlit~=1.22.0